fix upstart script
When the OTP application was being booted on net-device-up I could never connect to the running BEAM.
I would have to manually restart the application via ssh and then I would be able to connect.
Switching to `start on runlevel[2345]` has resolved the issue (for reasons I do not understand).